blocks/userStoryCompiler/requirements.txt
PyYAML
nltk

fastapi
uvicorn[standard]
